Create PaintAreaApp
Week 1:
Bring in user input, use in calculations, and provide output. 
This application takes in a diameter of a field and a desired paint color from the user and prints how many buckets of paint will be needed to cover the field. Each color has a different rate of coverage.